requirements.txt
Flask==3.1.2
firebase-admin==7.1.0

